requirements.txt
django
djangorestframework
ipdb
django-angular
